Add basic configuration for OSL mysql servers.
This enables the Percona repository, the Percona MySQL server and
client, and sets necessary configuration to prevent the service from
cleanly restarting (a side effect of the Percona rpm automatically
starting mysqld when installed).

This is the beginning of something great and will let us move away from
the previous forked mysql cookbook we are currently using.