Include additional GPG key for RPMs for Percona
It seems as though the latest builds of the percona RPMs including signing with
an additional key which is not listed in the yum repository. This is breaking
automatic yum upgrades. I've created an upstream issue with Percona
percona/percona-repositories#3 to see if they can get this published on their
website. Afterwards, I'll submit a PR to the sous-chefs/percona repository.