Add support for installing Percona client in osl-mysql::client
Currently osl-mysql::client assumes you're not using Percona. This becomes a
problem especially in a testing environment where we may install Percona server
but still need to have the client installed. To work around this, I've made the
following changes:

- Created a new attribute node['osl-mysql']['enable_percona_client'] (false by
  default)
- When using Percona, use percona::client instead of the mysql_client resource
  for better portability.
- Ensure we also install the devel packages for all variants of mysql

This should not impact current installations AFAIK other than maybe installing
the devel package if it doesn't already have it.